Restructure kubevirt file layout
