+++ robots/cmd/kubevirt/BUILD.bazel
@@ -5,7 +5,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "kubevirt.io/project-infra/robots/cmd/kubevirt",
     visibility = ["//visibility:private"],
-    deps = ["//robots/pkg/kubevirt:go_default_library"],
+    deps = ["//robots/pkg/kubevirt/cmd:go_default_library"],
 )
 
 go_binary(
