+++ robots/cmd/kubevirt/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
-	"kubevirt.io/project-infra/robots/pkg/kubevirt"
+	"kubevirt.io/project-infra/robots/pkg/kubevirt/cmd"
 	"os"
 )
 
 func main() {
-	kubevirt.Execute()
+	cmd.Execute()
 	os.Exit(0)
 }
