--- robots/pkg/kubevirt/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["root.go"],
-    importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//robots/pkg/kubevirt/copy:go_default_library",
-        "//robots/pkg/kubevirt/flags:go_default_library",
-        "//robots/pkg/kubevirt/require:go_default_library",
-        "@com_github_spf13_cobra//:go_default_library",
-    ],
-)
