+++ robots/pkg/kubevirt/cmd/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["root.go"],
+    importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt/cmd",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//robots/pkg/kubevirt/cmd/copy:go_default_library",
+        "//robots/pkg/kubevirt/cmd/flags:go_default_library",
+        "//robots/pkg/kubevirt/cmd/require:go_default_library",
+        "@com_github_spf13_cobra//:go_default_library",
+    ],
+)
