+++ robots/pkg/kubevirt/cmd/copy/copy.go
