+++ robots/pkg/kubevirt/cmd/flags/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["flags.go"],
-    importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt/flags",
+    importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt/cmd/flags",
     visibility = ["//visibility:public"],
     deps = [
         "//robots/pkg/kubevirt/log:go_default_library",
