+++ robots/pkg/kubevirt/github/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "kubevirt.io/project-infra/robots/pkg/kubevirt/github",
     visibility = ["//visibility:public"],
     deps = [
-        "//robots/pkg/kubevirt/flags:go_default_library",
+        "//robots/pkg/kubevirt/cmd/flags:go_default_library",
         "//robots/pkg/kubevirt/log:go_default_library",
         "@com_github_google_go_github//github:go_default_library",
         "@org_golang_x_oauth2//:go_default_library",
